# nodejs-server-server/api/swagger.yaml
...
swagger: "2.0"
info:
  version: "0.0.1"
  title: "Home Iot App"
host: "localhost:10010"
basePath: "/"
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /currenttime:
    get:
      description: "Returns the current time in the requested format"
      operationId: "currenttime"
      parameters:
      - name: "format"
        in: "query"
        description: "The format in which the caller wants to have the current time"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/HelloWorldResponse"
        default:
          description: "Error"
          schema:
            $ref: "#/definitions/ErrorResponse"
    x-swagger-router-controller: "hello_world"
...
definitions:
  HelloWorldResponse:
    required:
    - "message"
    properties:
      message:
        type: "string"
    example:
      message: "message"
...
  ErrorResponse:
    required:
    - "message"
    properties:
      message:
        type: "string"